+++ infra/modules/azure_function_app/outputs.tf
@@ -1,7 +1,7 @@
 output "subnet" {
   value = {
-    id   = azurerm_subnet.this.id
-    name = azurerm_subnet.this.name
+    id   = try(azurerm_subnet.this[0].id, var.subnet_id)
+    name = try(azurerm_subnet.this[0].name, null)
   }
 }
 
@@ -16,7 +16,7 @@ output "function_app" {
   value = {
     resource_group_name = azurerm_linux_function_app.this.resource_group_name
     plan = {
-      id   = try(azurerm_service_plan.this[0].id, null)
+      id   = try(azurerm_service_plan.this[0].id, var.app_service_plan_id)
       name = try(azurerm_service_plan.this[0].name, null)
     }
     function_app = {