+++ infra/modules/azure_function_app/subnets.tf
@@ -1,4 +1,6 @@
 resource "azurerm_subnet" "this" {
+  count = local.function_app.has_existing_subnet ? 0 : 1
+
   name                 = local.subnet.name
   virtual_network_name = data.azurerm_virtual_network.this.name
   resource_group_name  = data.azurerm_virtual_network.this.resource_group_name