+++ infra/modules/azure_role_assignments/CHANGELOG.md
@@ -1,5 +1,17 @@
 # azure_role_assignments
 
+## 0.1.2
+
+### Patch Changes
+
+- 075f30a: Fixed roles for API Management
+
+## 0.1.1
+
+### Patch Changes
+
+- 8dda982: Add a description in the package.json file
+
 ## 0.1.0
 
 ### Minor Changes