+++ infra/scripts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # pre_commit_scripts
 
+## 0.0.2
+
+### Patch Changes
+
+- 8dda982: Add a description in the package.json file
+
 ## 0.0.1
 
 ### Patch Changes